Fix duplicate extended error-code assignment path

+++ CiccioSoft.Data.Sqlite/SqliteTransaction.cs
@@ -96,7 +96,7 @@ public class SqliteTransaction : DbTransaction
         }
         catch (SqliteInteropException ex)
         {
-            throw new SqliteException(ex.Message, (int)ex.BaseErrorCode, ex.ExtendedErrorCode, ex);
+            throw new SqliteException(ex.Message, (int)ex.BaseErrorCode, (int)ex.ExtendedErrorCode, ex);
         }
         finally
         {
